jaundice_ml/requirements.txt
numpy>=1.21.0
pandas>=1.5.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scikit-learn>=1.1.0
xgboost>=1.6.0
joblib>=1.1.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
Pillow>=9.0.0
requests>=2.25.0
